requirements.txt
Pillow==9.4.0
pandas==1.5.3